Implement POST /beds/search endpoint
